requirements.txt
aiohttp>=3.9
cryptography>=42.0
python-dotenv>=1.0
requests>=2.31
